requirements.txt
streamlit>=1.35.0
google-genai>=0.8.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
supabase>=2.0.0
